Extract worker threads to src/gui/workers.py
Lifts the QThread workers, the shared progress dialog, and the cell-edit
delegate out of main_window.py so each can be reasoned about (and later
tested) in isolation. Net: main_window.py drops from 3900 to 3521 lines.

New modules
-----------
- src/gui/workers.py (402 lines):
    AnimatedProgressDialog      — indeterminate↔determinate progress dialog
    FileLoaderWorker            — load sources, build StringEntry list
    StartupSyncWorker           — refresh URL-backed sources on startup
    EnhancementsGeneratorWorker — run scripts/generate_enhancements_ini.py
    P4kExtractWorker            — unp4k extraction of global.ini
    DataForgeExtractWorker      — unp4k + unforge + patch pipeline
    SelectAllDelegate           — Custom Value cell delegate (auto-select +
                                  EM3/EM4 wrap actions)

- src/utils/resource_path.py (34 lines):
    get_resource_path(rel) — PyInstaller-aware bundled-resource resolver
    resolve_patches_dir()  — Path → bundled DataForge patches dir
  Lives in src/utils/ rather than src/gui/ because the workers and the
  main window both depend on it; keeping it leaf-level avoids workers
  importing from main_window or vice versa.

Behavior preserved exactly — same class bodies, same signal contracts,
same in-method imports. The only renames:
- get_resource_path: argument was untyped, now str (was already a str
  at every call site).
- _resolve_patches_dir → resolve_patches_dir (dropped the leading
  underscore: it's now a public utility, no longer module-private).

main_window.py changes
----------------------
- Removed the 7 extracted classes and the two resource-path helpers.
- Consolidated the mid-file imports (was ad-hoc inserted at line 100
  between AnimatedProgressDialog and the rest of the file) up to the
  top alphabetically.
- New imports: workers (the 7 names) and resource_path.get_resource_path.

AppUpdateCheckWorker stays in src/utils/app_updater.py — unchanged.
This PR only reorganises what was already in main_window.py.

New tests
---------
tests/test_resource_path.py — 9 tests covering get_resource_path
(unfrozen / frozen via _MEIPASS / nested paths / monkeypatch hygiene)
and resolve_patches_dir. Pure-function tests — no Qt required.

Worker tests are out of scope: they need pytest-qt (not currently a
dev dependency); adding it is a separate decision.

Test result
-----------
124 pass (was 115 baseline), 4 pak_extraction failures unchanged
(unrelated pre-existing dataforge_cache_dir str/Path bug). test_core.py
still doesn't collect on this branch — that's PR #6's territory.

Module sources adapted from jonigirl's work in PR #5; the resource_path
extraction, the import consolidation in main_window.py, and the test
coverage are this PR's additions.
